requirements.txt
# Smart Panel v2.0 Requirements
# Core functionality
gpiozero>=1.6.0
luma.lcd>=2.9.0
Pillow>=8.0.0
psutil>=5.8.0
RPi.GPIO>=0.7.0

# Note: lgpio and spidev are installed as system packages
# They will be available via --system-site-packages flag

# Matter IoT support (optional but recommended)
# chip-matter>=0.0.1  # Uncomment when available

# QR Code generation for Matter commissioning
qrcode[pil]>=7.0

# Development and testing (optional)
# pytest>=6.0.0  # Uncomment for testing
# flake8>=3.9.0  # Uncomment for linting
